fix: existing tests after refactor
