+++ client/src/components/geneExpression/infoPanel/common/infoPanelParts/index.tsx
