+++ client/src/components/geneExpression/infoPanel/infoPanelCellType/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { connect } from "react-redux";
 import { Props, mapStateToProps } from "./types";
-import ContainerInfo from "../common/containerInfo";
+import ContainerInfo from "../common/infoPanelContainer";
 import { CELLGUIDE_URL } from "../common/constants";
 
 function CellTypeInfo(props: Props) {
