+++ client/src/components/geneExpression/infoPanel/infoPanelCellType/types.ts
