+++ client/src/components/geneExpression/infoPanel/infoPanelGene/types.ts
