+++ server/common/utils/cell_type_info.py
@@ -1,5 +1,6 @@
-import requests
 from typing import Dict, List, TypedDict, cast
+
+import requests
 
 from server.common.constants import CELLGUIDE_BASE_URL
